+++ spec/cerbos/client_spec.rb
@@ -42,7 +42,14 @@ RSpec.describe Cerbos::Client do
               token: JWT.encode({delete: true}, nil, "none")
             }
           },
-          request_id: "42"
+          request_id: "42",
+          request_context: cerbos_version_at_least?("0.51.0") ? {
+            annotations: {
+              foo: {
+                bar: "baz"
+              }
+            }
+          } : nil
         )
       end
 
@@ -104,7 +111,14 @@ RSpec.describe Cerbos::Client do
             }
           },
           include_metadata: true,
-          request_id: "42"
+          request_id: "42",
+          request_context: cerbos_version_at_least?("0.51.0") ? {
+            annotations: {
+              foo: {
+                bar: "baz"
+              }
+            }
+          } : nil
         )
       end
 
@@ -149,6 +163,7 @@ RSpec.describe Cerbos::Client do
             if cerbos_version_at_least?("0.27.0")
               [
                 Cerbos::Output::CheckResources::Result::Output.new(
+                  action: cerbos_version_at_least?("0.51.0") ? "delete" : "",
                   source: "resource.document.v1#delete",
                   value: "delete_allowed:me@example.com"
                 )
@@ -219,7 +234,14 @@ RSpec.describe Cerbos::Client do
             }
           },
           include_metadata: true,
-          request_id: "42"
+          request_id: "42",
+          request_context: cerbos_version_at_least?("0.51.0") ? {
+            annotations: {
+              foo: {
+                bar: "baz"
+              }
+            }
+          } : nil
         )
       end
 
@@ -267,6 +289,7 @@ RSpec.describe Cerbos::Client do
                 if cerbos_version_at_least?("0.27.0")
                   [
                     Cerbos::Output::CheckResources::Result::Output.new(
+                      action: cerbos_version_at_least?("0.51.0") ? "delete" : "",
                       source: "resource.document.v1#delete",
                       value: "delete_allowed:me@example.com"
                     )
@@ -315,6 +338,7 @@ RSpec.describe Cerbos::Client do
                 if cerbos_version_at_least?("0.27.0")
                   [
                     Cerbos::Output::CheckResources::Result::Output.new(
+                      action: cerbos_version_at_least?("0.51.0") ? "delete" : "",
                       source: "resource.document.v1#delete",
                       value: "delete_allowed:me@example.com"
                     )
@@ -368,6 +392,7 @@ RSpec.describe Cerbos::Client do
                 if cerbos_version_at_least?("0.27.0")
                   [
                     Cerbos::Output::CheckResources::Result::Output.new(
+                      action: cerbos_version_at_least?("0.51.0") ? "delete" : "",
                       source: "resource.document.v1#delete",
                       value: "delete_allowed:me@example.com"
                     )
@@ -410,7 +435,14 @@ RSpec.describe Cerbos::Client do
               }
             },
             include_metadata: true,
-            request_id: "42"
+            request_id: "42",
+            request_context: cerbos_version_at_least?("0.51.0") ? {
+              annotations: {
+                foo: {
+                  bar: "baz"
+                }
+              }
+            } : nil
           )
         end
 
@@ -479,7 +511,14 @@ RSpec.describe Cerbos::Client do
               }
             },
             include_metadata: true,
-            request_id: "42"
+            request_id: "42",
+            request_context: cerbos_version_at_least?("0.51.0") ? {
+              annotations: {
+                foo: {
+                  bar: "baz"
+                }
+              }
+            } : nil
           )
         end